move lifecycle rules (which is just the init handler) to separate file
[ci skip-rust]

[ci skip-build-wheels]
